+++ Coimbra/Events/FixedUpdateEvent.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine.Scripting;
+
+namespace Coimbra
+{
+    [Preserve]
+    public readonly struct FixedUpdateEvent
+    {
+        public readonly float DeltaTime;
+
+        public FixedUpdateEvent(float deltaTime)
+        {
+            DeltaTime = deltaTime;
+        }
+    }
+}