+++ Coimbra/IService.cs
@@ -7,5 +7,12 @@ namespace Coimbra
     /// Base interface for any service to be used with <see cref="ServiceLocator"/>.
     /// </summary>
     [RequireImplementors]
-    public interface IService : IDisposable { }
+    public interface IService : IDisposable
+    {
+        /// <summary>
+        /// The <see cref="ServiceLocator"/> that owns this service.
+        /// <para></para><see cref="set_OwningLocator"/> is for internal use only.
+        /// </summary>
+        ServiceLocator OwningLocator { get; set; }
+    }
 }