+++ Coimbra/Services/IApplicationService.cs
@@ -1,0 +1,7 @@
+﻿namespace Coimbra
+{
+    /// <summary>
+    /// Responsible for the application lifetime cycle, meant to be used to fire Unity's callbacks events.
+    /// </summary>
+    public interface IApplicationService : IService { }
+}